+++ phb-app/src/pages/Armor.vue
@@ -20,45 +20,93 @@
           But this doesn't mean a Drell Soldier can't strap on some heavy armor. It means you'll
           have to take Feats to get yourself proficient in heavy armor. Or you can modify your light armor to have more AC.
     rule-card
-      span(slot="header") Armor isn't ALL IN ONE
+      span(slot="header") Armor isn't ALL-IN-ONE
       template(slot="text")
         p.
-          In ME5e there are 4 armor placements, Head, Chest, Arms, Legsand you can mix and match armor types (Light, Medium, Heavy).
-          This makes the array of AC possibilities a bit more complex. Remember, you must be proficient in an armor piece's type to wear it.
-        p.headline AC Breakdown
+          In Mass Effect 5e, there are 4 armor placements: #[strong Head], #[strong Chest], #[strong Arms], and #[strong Legs].
+          You can mix and match armor types (#[strong Light], #[strong Medium], #[strong Heavy]). This makes the array of
+          Armor Class (AC) possibilities a bit more complex.
+        p.headline.mb-0 AC Breakdown
         table.table
           thead.text-xs-left
             tr
               th Armor Type Mix
               th AC
+              th Min STR
+              th Stealth
           tbody
             tr
               td Empty armor slot
               td -1 AC
+              td
+              td
             tr
               td All Light
               td 11 + Dexterity Modifier
+              td -
+              td -
             tr
               td Medium & Light
               td 12 + Dexterity Modifier (max 2)
+              td 8
+              td -
             tr
               td All Medium
               td 13 + Dexterity Modifier (max 2)
+              td 10
+              td -
             tr
               td A least 1 Heavy, Mix of Medium & Light
               td 14
+              td 11
+              td Disadvantage
             tr
               td 1 Heavy, All other Medium
               td 15
+              td 12
+              td Disadvantage
             tr
               td 2 Heavy & 2 Medium
               td 16
+              td 13
+              td Disadvantage
             tr
               td 3 Heavy & 1 Medium
               td 17
+              td 14
+              td Disadvantage
             tr
               td All Heavy
               td 18
+              td 15
+              td Disadvantage
+        v-alert(type="warning" v-bind:value="true").
+          You must be proficient in an armor type (Light, Medium, or Heavy) to wear it.
+        p.headline.mt-3 Variant: Encumbrance
+        p.mb-0 If using the Encumbrance rule, use the following to calculate your armor weight.
+        table.table
+            thead.text-xs-left
+              tr
+                th
+                th Head
+                th Chest
+                th Arms & Legs
+            tbody
+              tr
+                td #[strong Light]
+                td 1
+                td 2
+                td 0
+              tr
+                td #[strong Medium]
+                td 1
+                td 3
+                td 1
+              tr
+                td #[strong Heavy]
+                td 2
+                td 4
+                td 2
     rule-card
       span(slot="header") Armor can be customized & created
       template(slot="text")
@@ -81,28 +129,29 @@
               th Standard Races
           tbody
             tr
-              td Light Armor
+              td #[strong Light]
               td 3,000
               td 8,000
               td 2,000
               td All races except, Geth, Krogan, and Turians
             tr
-              td Medium Armor
+              td #[strong Medium]
               td 4,000
               td 10,000
               td 3,000
               td Asari, Batarian, Human, Quarian, Turian, Volus
             tr
-              td Heavy Armor
+              td #[strong Heavy]
               td 5,000
               td 12,000
               td 4,000
               td Batarian, Human, Krogan, Turian
             tr
-              td Mod Slots (Max)
-              td 1 (max 3)
-              td 2 (max 3)
-              td 0 (max 2)
+              td Mod Slots #[em (max)]
+              td 1 #[em (3)]
+              td 2 #[em (3)]
+              td 0 #[em (2)]
+              td
         ul.list
           li #[strong Additional Mod Slots:] 20,000 credits
           li #[strong Race Custom Fitting:] 5,000 credits
@@ -110,30 +159,27 @@
           v-btn(to="/armor/mods") Mod List
         div.hr
         p.headline Example
-        ol.list
-          li.
-            You want a Medium-Type, Head Armor for your Drell. On the table, the base cose for Medium + Head is 4,000 credits.
-          li.
-            But this is for a Drell which is not a standard race for medium armor. So it will cost an additional 5,000 credits to get
-            this customized.
-          li.
-            You want to add 2 mods to the helmet, but a Head piece only starts with one mod slot. It will cost another 20,000 credits to upgrade
-            the helmet.
-          li.
-            Finally, you can add your mods (link above). You select the Umbra AI (8,000) and the Voice
-            Modulator (2,000).
-          li.
-            So the grand total for your custom head piece is: 39,000 credits.
-        p.headline.mt-3 Customizing Armor
+        p Medium Head Armor w/ 2 mod slots for a Drell
+        table.table
+          tbody
+            tr
+              td Medium Head
+              td 4000
+            tr
+              td Additional Mod Slot (starts with 1)
+              td 20000
+            tr
+              td Customized for Drell (not a standard race for Medium armor)
+              td 5000
+            tr
+              td #[strong Total]
+              td #[strong 29000 credits]
+        p.headline.mt-3 Looted armor
         p.
-          Instead, you might find it cheaper to take what you can get from the battlefield and customize it. If the in this example Drell got
-          a Medium Head piece off the battlefield with 2 mod slots, he could spend 15,000 credits to add the 2 mods and customize it
-          for his race.
-        p
-          strong GM Note:
-          span.
-            When awarding Armor, be sure to say the number of mod slots it came with. Only premium quality armor would have
-            more that the starting amount in the table above. For really cheap armor, you can remove mod slots.
+          You can also apply custom mods, additional mod slots, and race customizations to armor you've looted from the battlefield. The same prices apply.
+        v-alert(type="info" v-bind:value="true").
+          #[strong GM Note]: When awarding Armor, be sure to say the number of mod slots it came with. Only premium quality armor would have
+          more than the starting amount in the table above. For really cheap armor, you can remove mod slots completely.
 </template>
 
 <script>